% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_json_for_shinysurvey.R
\name{convert_questions_logics_to_df}
\alias{convert_questions_logics_to_df}
\title{Format survey questions from json file into data frame required for \code{shinysurveys}.}
\usage{
convert_questions_logics_to_df(columns_to_numeric)
}
\description{
Format survey questions from json file into data frame required for \code{shinysurveys}.
}
