% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_json_for_shinysurvey.R
\name{convert_questions_logics_to_results}
\alias{convert_questions_logics_to_results}
\title{Format survey results from json file into data frame.}
\usage{
convert_questions_logics_to_results(columns_to_numeric)
}
\description{
Format survey results from json file into data frame.
}
