% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_json_for_shinysurvey.R
\name{format_dependencies_logic}
\alias{format_dependencies_logic}
\title{Given an input data frame formatted with the columns names required for \code{shinysurvey}, add
questions dependencies.}
\usage{
format_dependencies_logic(survey_qns)
}
\description{
Given an input data frame formatted with the columns names required for \code{shinysurvey}, add
questions dependencies.
}
