% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_json_for_shinysurvey.R
\name{import_json_for_shinysurvey}
\alias{import_json_for_shinysurvey}
\title{Load file with risk assessment logics from https://github.com/summerscope/summerchildpy/blob/main/questions.json
and parse into (1) the input format required for \code{shinysurveys} and the survey result categorisation.}
\usage{
import_json_for_shinysurvey()
}
\description{
Load file with risk assessment logics from https://github.com/summerscope/summerchildpy/blob/main/questions.json
and parse into (1) the input format required for \code{shinysurveys} and the survey result categorisation.
}
