% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_json_for_shinysurvey.R
\name{import_questions_logics}
\alias{import_questions_logics}
\title{Load file with risk assessment logics from https://github.com/summerscope/summerchildpy/blob/main/questions.json}
\usage{
import_questions_logics()
}
\description{
Load file with risk assessment logics from https://github.com/summerscope/summerchildpy/blob/main/questions.json
}
