% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_json_for_shinysurvey.R
\name{import_questions_logics_to_df}
\alias{import_questions_logics_to_df}
\title{Load file with risk assessment logics from https://github.com/summerscope/summerchildpy/blob/main/questions.json
and format into dataframe required for \code{shinysurveys}.}
\usage{
import_questions_logics_to_df()
}
\description{
Load file with risk assessment logics from https://github.com/summerscope/summerchildpy/blob/main/questions.json
and format into dataframe required for \code{shinysurveys}.
}
