% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSummerChildApp.R
\name{runSummerChildApp}
\alias{runSummerChildApp}
\title{Load questions and results categorisation from JSON file, run survey and then present results and recommendations based on survey responses}
\usage{
runSummerChildApp()
}
\description{
Load questions and results categorisation from JSON file, run survey and then present results and recommendations based on survey responses
}
